refactored NameManager logic for new design
